keymap: add vial_secure + better support for user keymaps + clean house.
keymap.c has been split into a stub keymap.c that contains the
keymap and keymap_support.c that actually has all the pointer logic
and other fun things that make Svalboard, Svalboard.

Moved .vil files because they didn't make sense in the vial directory
or the vial_secure directory.

Changed the default for vial back to insecure.  With the new
keymap, users can easily build firmware if needed, if there is
enough demand we can even provide vial_secure firmware.

Keymaps that were not in use were removed.
