+++ keyboards/svalboard/keymaps/README.md
@@ -1,0 +1,11 @@
+Keymaps Information:
+
+Svalboard keeps keymap_support.c up to date, with the most recent developments
+on the board.  If you want to see an example keymap, look at the vial keymap.
+
+You do not have to use vial with Svalboard, but the vial, and vial_secure
+keymaps will be kept up to date, others may be out of date.
+
+The API between Svalboard and a keymap may change over time.  Please feel
+free to talk to us on Discord about changes you may need, or that have
+impacted you.
