--- keyboards/svalboard/keymaps/claussen/vial.json
@@ -1,242 +1,0 @@
-{
-    "name": "Svalboard Vial Definition",
-    "author": "Morgan Venable",
-    "notes": "This is a Vial Keymap template for Svalboard",
-    "vendorId":"0x303A",
-    "productId":"0x4044",
-    "lighting": "none",
-    "matrix": {
-        "rows": 10,
-        "cols": 6
-    },
-    "layouts": {
-           "keymap": [
-            [
-              {
-                "x": 6,
-                "c": "#3f3f3f",
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00",
-                "fa": [
-                  7,
-                  0,
-                  0,
-                  0
-                ]
-              },
-              "2,3",
-              {
-                "x": 7.5
-              },
-              "7,3"
-            ],
-            [
-              {
-                "x": 5,
-                "t": "#ffffff\n#3b64ff"
-              },
-              "2,4",
-              {
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00"
-              },
-              "2,2",
-              {
-                "t": "#ffffff\n#3b64ff"
-              },
-              "2,1",
-              {
-                "x": 5.5,
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00"
-              },
-              "7,4",
-              "7,2",
-              "7,1"
-            ],
-            [
-              {
-                "y": -0.5,
-                "x": 3.5
-              },
-              "3,3",
-              {
-                "x": 4
-              },
-              "1,3",
-              {
-                "x": 2.5
-              },
-              "6,3",
-              {
-                "x": 4
-              },
-              "8,3"
-            ],
-            [
-              {
-                "y": -0.5,
-                "x": 6
-              },
-              "2,0",
-              {
-                "x": 7.5
-              },
-              "7,0"
-            ],
-            [
-              {
-                "y": -0.5,
-                "x": 2.5,
-                "t": "#ffffff"
-              },
-              "3,4",
-              {
-                "t": "#ffffff\n#3b64ff"
-              },
-              "3,2",
-              {
-                "t": "#ffffff\n\n\n#f9ff00"
-              },
-              "3,1",
-              {
-                "x": 2
-              },
-              "1,4",
-              {
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00"
-              },
-              "1,2",
-              "1,1",
-              {
-                "x": 0.5
-              },
-              "6,4",
-              "6,2",
-              "6,1",
-              {
-                "x": 2
-              },
-              "8,4",
-              "8,2",
-              "8,1"
-            ],
-            [
-              {
-                "y": -0.5,
-                "x": 1
-              },
-              "4,3",
-              {
-                "x": 17.5
-              },
-              "9,3"
-            ],
-            [
-              {
-                "y": -0.5,
-                "x": 3.5
-              },
-              "3,0",
-              {
-                "x": 4
-              },
-              "1,0",
-              {
-                "x": 2.5
-              },
-              "6,0",
-              {
-                "x": 4
-              },
-              "8,0"
-            ],
-            [
-              {
-                "y": -0.5,
-                "t": "#ffffff"
-              },
-              "4,4",
-              {
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00"
-              },
-              "4,2",
-              "4,1",
-              {
-                "x": 15.5
-              },
-              "9,4",
-              "9,2",
-              {
-                "t": "#ffffff\n\n\n#f9ff00"
-              },
-              "9,1"
-            ],
-            [
-              {
-                "x": 1,
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00"
-              },
-              "4,0",
-              {
-                "x": 5.25,
-                "t": "#ffffff"
-              },
-              "0,1",
-              {
-                "x": 1
-              },
-              "0,4",
-              {
-                "x": 1
-              },
-              "5,4",
-              {
-                "x": 1
-              },
-              "5,1",
-              {
-                "x": 5.25,
-                "t": "#ffffff\n#3b64ff\n\n#f9ff00"
-              },
-              "9,0"
-            ],
-            [
-              {
-                "x": 8.25,
-                "t": "#ffffff"
-              },
-              "0,2",
-              {
-                "x": 3
-              },
-              "5,2"
-            ],
-            [
-              {
-                "x": 8.25
-              },
-              "0,5",
-              {
-                "x": 3
-              },
-              "5,5"
-            ],
-            [
-              {
-                "x": 7.25
-              },
-              "0,0",
-              {
-                "x": 1
-              },
-              "0,3",
-              {
-                "x": 1
-              },
-              "5,3",
-              {
-                "x": 1
-              },
-              "5,0"
-            ]
-        ]
-    }
-}
