--- keyboards/svalboard/keymaps/nocodes/readme.MD
@@ -1,5 +1,0 @@
-This is a codeless, 5 layer keymap compatible with loading .vil files to set everything up.
-
-It lets you test the entire keyboard in the Vial matrix tester without worrying about spurious keys due to keys being not installed or hardware being bad.
-
-To build it, simply copy it over keymap.c in the vial folder.  Remember to save the old one if you're faffing about with this...
