+++ keyboards/svalboard/keymaps/vial/keymap.c
@@ -15,42 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include QMK_KEYBOARD_H
-#include <stdbool.h>
-#include <stdint.h>
-#include "svalboard.h"
-
-#define MH_AUTO_BUTTONS_LAYER MBO
-#define MH_AUTO_BUTTONS_TIMEOUT 5000
-#define PS2_MOUSE_SCROLL_BTN_MASK (1<<PS2_MOUSE_BTN_MIDDLE) // this mask disables the key for non-PS2 purposes
-
-void keyboard_post_init_user(void) {
-  // Customise these values if you need to debug the matrix
-  //debug_enable=true;
-  debug_matrix=true;
-  //debug_keyboard=true;
-  //debug_mouse=true;
-}
-
-// in keymap.c:
-#ifdef POINTING_DEVICE_AUTO_MOUSE_ENABLE
-void pointing_device_init_user(void) {
-    set_auto_mouse_layer(15); // only required if AUTO_MOUSE_DEFAULT_LAYER is not set to index of <mouse_layer>
-    set_auto_mouse_enable(true);         // always required before the auto mouse feature will work
-}
-#endif
-
-enum my_keycodes {
-  SV_LEFT_DPI_INC = QK_KB_0,
-  SV_LEFT_DPI_DEC,
-  SV_RIGHT_DPI_INC,
-  SV_RIGHT_DPI_DEC,
-  SV_LEFT_SCROLL_TOGGLE,
-  SV_RIGHT_SCROLL_TOGGLE,
-  SV_RECALIBRATE_POINTER,
-  KC_NORMAL_HOLD = SAFE_RANGE,
-  KC_FUNC_HOLD,
-};
+#include "../keymap_support.c"
 
 enum layer {
     NORMAL,
@@ -58,24 +23,13 @@ enum layer {
     FUNC,
     FUNC_HOLD,
     NAS,
-    L5,
-    L6,
-    L7,
-    L8,
-    L9,
-    L10,
-    L11,
-    L12,
-    L13,
-    L14,
-    MBO,
-    NUM_LAYERS
+    MBO = MH_AUTO_BUTTONS_LAYER,
 };
 
-const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
- [NORMAL] = LAYOUT(
-             /*Center           North           East            South           West*/
-        
+const uint16_t PROGMEM keymaps[DYNAMIC_KEYMAP_LAYER_COUNT][MATRIX_ROWS][MATRIX_COLS] = {
+    [NORMAL] = LAYOUT(
+        /*Center           North           East            South           West*/
+
         /*R1*/ KC_J,            KC_U,           KC_QUOTE,       KC_M,           KC_H,
         /*R2*/ KC_K,            KC_I,           KC_COLON,       KC_COMMA,       KC_Y,
         /*R3*/ KC_L,            KC_O,           KC_LGUI,        KC_DOT,         KC_N,
@@ -86,13 +40,13 @@ const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
         /*L3*/ KC_S,            KC_W,           KC_B,           KC_X,           KC_ESC,
         /*L4*/ KC_A,            KC_Q,           KC_LBRC,        KC_Z,           KC_DEL,
 
-        /*Down                  Inner (pad)     Upper (Mode)    O.Upper (nail)  OL (knuckle) Pushthrough*/  
+        /*Down                  Inner (pad)     Upper (Mode)    O.Upper (nail)  OL (knuckle) Pushthrough*/
         /*RT*/ MO(NAS),         KC_SPACE,       TO(FUNC),       KC_BSPC,        KC_LALT,     TG(NAS),
         /*LT*/ KC_LSFT,         KC_ENTER,       TO(NORMAL),          KC_TAB,         KC_LCTL,     KC_CAPS
-    ),
+        ),
 
     [NORMAL_HOLD] = LAYOUT(
-             /*Center           North           East            South           West*/
+        /*Center           North           East            South           West*/
         /*R1*/ KC_LEFT,         KC_WH_L,        XXXXXXX,        KC_MS_L,        LCTL(KC_LEFT),
         /*R2*/ KC_DOWN,         KC_WH_D,        XXXXXXX,        KC_MS_D,        LCTL(KC_DOWN),
         /*R3*/ KC_UP,           KC_WH_U,        XXXXXXX,        KC_MS_U,        LCTL(KC_UP),
@@ -106,10 +60,10 @@ const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
         /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/
         /*RT*/ _______,         _______,        _______,        _______,        _______, _______,
         /*LT*/ _______,         _______,        _______,        _______,        _______, _______
-    ),
+        ),
 
     [FUNC] = LAYOUT(
-             /*Center           North           East            South           West*/
+        /*Center           North           East            South           West*/
         /*R1*/ KC_HOME,         KC_UP,          KC_RIGHT,       KC_DOWN,        KC_LEFT,
         /*R2*/ XXXXXXX,         KC_F8,          XXXXXXX,        KC_F7,          KC_END,
         /*R3*/ KC_PSCR,         KC_F10,         KC_LGUI,        KC_F9,          KC_INS,
@@ -120,13 +74,13 @@ const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
         /*L3*/ XXXXXXX,         KC_F4,          XXXXXXX,        KC_F3,          KC_ESC,
         /*L4*/ XXXXXXX,         KC_F2,          XXXXXXX,        KC_F1,          KC_DEL,
 
-             /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/  
+        /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/
         /*RT*/ MO(NAS),         KC_SPACE,       _______,       KC_BSPC,      KC_LALT, _______,
         /*LT*/ KC_LSFT,       KC_ENTER,         _______, KC_TAB,         KC_LCTL,_______
-    ),
+        ),
 
     [FUNC_HOLD] = LAYOUT(
-             /*Center           North           East            South           West*/
+        /*Center           North           East            South           West*/
         /*R1*/ KC_LEFT,         LCTL(KC_UP),   LCTL(KC_RIGHT), LCTL(KC_DOWN), LCTL(KC_LEFT),
         /*R2*/ KC_UP,           KC_MS_U,        KC_MS_R,        KC_MS_D,        KC_MS_L,
         /*R3*/ KC_DOWN,         KC_WH_U,        KC_WH_R,        KC_WH_D,        KC_WH_L,
@@ -137,13 +91,13 @@ const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
         /*L3*/ XXXXXXX,         XXXXXXX,        XXXXXXX,        XXXXXXX,     XXXXXXX,
         /*L4*/ _______,      _______,        _______,        _______,       _______,
 
-             /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/  
+        /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/
         /*RT*/ _______,         _______,        _______,        _______,        _______,_______,
         /*LT*/ _______,         _______,        _______,        _______,        _______, _______
-    ),
+        ),
 
     [NAS] = LAYOUT(
-             /*Center           North           East            South           West*/
+        /*Center           North           East            South           West*/
         /*R1*/ KC_7,            KC_AMPR,        KC_UNDS,        KC_KP_PLUS,     KC_6,
         /*R2*/ KC_8,            KC_KP_ASTERISK, KC_COLON,       KC_COMMA,       KC_CIRCUMFLEX,
         /*R3*/ KC_9,            KC_LPRN,        KC_LGUI,        KC_DOT,         KC_SEMICOLON,
@@ -154,13 +108,13 @@ const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
         /*L3*/ KC_2,            KC_AT,          XXXXXXX,        KC_X,           KC_ESC,
         /*L4*/ KC_1,            KC_EXCLAIM,     KC_TILDE,       KC_EQUAL,       KC_DEL,
 
-        /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/  
+        /*Down                  Inner           Upper           Outer Upper     Outer Lower  Pushthrough*/
         /*RT*/ MO(NAS),         KC_SPACE,       _______,       KC_BSPC,        KC_LALT, _______,
         /*LT*/ KC_LSFT,         KC_ENTER,       _______,        KC_TAB,         KC_LCTL, _______
-    ),
-    
+        ),
+
     [MBO] = LAYOUT(
-             /*Center           North           East            South           West*/
+        /*Center           North           East            South           West*/
         /*R1*/ KC_TRNS,        KC_TRNS,       KC_TRNS,       KC_BTN1,       KC_TRNS,
         /*R2*/ KC_TRNS,        KC_TRNS,       KC_TRNS,       KC_BTN3,       KC_TRNS,
         /*R3*/ KC_TRNS,        KC_TRNS,       KC_TRNS,       KC_BTN2,       KC_TRNS,
@@ -174,208 +128,3 @@ const uint16_t PROGMEM keymaps[NUM_LAYERS][MATRIX_ROWS][MATRIX_COLS] = {
         )
 
 };
-
-
-#if (defined MH_AUTO_BUTTONS && defined PS2_MOUSE_ENABLE && defined MOUSEKEY_ENABLE)  || defined(POINTING_DEVICE_AUTO_MOUSE_MH_ENABLE)
-
-static uint16_t mh_auto_buttons_timer;
-extern int tp_buttons; // mousekey button state set in action.c and used in ps2_mouse.c
-
-void mouse_mode(bool);
-
-#endif
-
-#if defined(POINTING_DEVICE_AUTO_MOUSE_MH_ENABLE)
-
-#define SCROLL_DIVISOR 20
-
-static int _ds_l_x = 0;
-static int _ds_l_y = 0;
-static int _ds_r_x = 0;
-static int _ds_r_y = 0;
-
-report_mouse_t pointing_device_task_combined_user(report_mouse_t reportMouse1, report_mouse_t reportMouse2) {
-    if (reportMouse1.x == 0 && reportMouse1.y == 0 && reportMouse2.x == 0 && reportMouse2.y == 0)
-        return pointing_device_combine_reports(reportMouse1, reportMouse2);
-
-    if (global_saved_values.left_scroll) {
-        int div_x;
-        int div_y;
-
-        _ds_l_x += reportMouse1.x;
-        _ds_l_y -= reportMouse1.y;
-
-        div_x = _ds_l_x / SCROLL_DIVISOR;
-        div_y = _ds_l_y / SCROLL_DIVISOR;
-        if (div_x != 0) {
-            reportMouse1.h += div_x;
-            _ds_l_x -= div_x * SCROLL_DIVISOR;
-        }
-
-        if  (div_y != 0) {
-            reportMouse1.v += div_y;
-            _ds_l_y -= div_y * SCROLL_DIVISOR;
-        }
-        reportMouse1.x = 0;
-        reportMouse1.y = 0;
-    }
-
-    if (global_saved_values.right_scroll) {
-        int div_x;
-        int div_y;
-
-        _ds_r_x += reportMouse2.x;
-        _ds_r_y -= reportMouse2.y;
-
-        div_x = _ds_r_x / SCROLL_DIVISOR;
-        div_y = _ds_r_y / SCROLL_DIVISOR;
-        if (div_x != 0) {
-            reportMouse2.h += div_x;
-            _ds_r_x -= div_x * SCROLL_DIVISOR;
-        }
-
-        if  (div_y != 0) {
-            reportMouse2.v += div_y;
-            _ds_r_y -= div_y * SCROLL_DIVISOR;
-        }
-        reportMouse2.x = 0;
-        reportMouse2.y = 0;
-    }
-
-    if (mh_auto_buttons_timer) {
-        mh_auto_buttons_timer = timer_read();
-    } else {
-        mouse_mode(true);
-#if defined CONSOLE_ENABLE
-        print("mh_auto_buttons: on\n");
-#endif
-    }
-    return pointing_device_combine_reports(reportMouse1, reportMouse2);
-}
-
-report_mouse_t pointing_device_task_user(report_mouse_t reportMouse) {
-    if (reportMouse.x == 0 && reportMouse.y == 0)
-        return reportMouse;
-
-    if (mh_auto_buttons_timer) {
-        mh_auto_buttons_timer = timer_read();
-    } else {
-        mouse_mode(true);
-#if defined CONSOLE_ENABLE
-        print("mh_auto_buttons: on\n");
-#endif
-    }
-    return reportMouse;
-}
-#endif
-
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-
-      // If console is enabled, it will print the matrix position and status of each key pressed
-#ifdef CONSOLE_ENABLE
-    uprintf("KL: kc: 0x%04X, col: %2u, row: %2u, pressed: %u, time: %5u, int: %u, count: %u\n", keycode, record->event.key.col, record->event.key.row, record->event.pressed, record->event.time, record->tap.interrupted, record->tap.count);
-#endif 
-
-#if (defined MH_AUTO_BUTTONS && defined PS2_MOUSE_ENABLE && defined MOUSEKEY_ENABLE) || defined(POINTING_DEVICE_AUTO_MOUSE_MH_ENABLE)
-    if (mh_auto_buttons_timer) {
-      switch (keycode) {
-      case KC_BTN1:
-      case KC_BTN2:
-      case KC_BTN3:
-      case KC_BTN4:
-      case KC_BTN5:
-      case KC_WH_U:
-      case KC_WH_D:
-      case KC_WH_R:
-      case KC_WH_L:
-      case KC_LSFT:
-      case KC_RSFT:
-      case KC_LCTL:
-      case KC_RCTL:
-      case KC_LALT:
-      case KC_RALT:
-      case KC_LGUI:
-      case KC_RGUI:
-      case SV_RECALIBRATE_POINTER:
-	break;
-      default:
-	mouse_mode(false);
-      }
-    }
-    if (!record->event.pressed) {
-       switch (keycode) {
-           case SV_LEFT_DPI_INC:
-                increase_left_dpi();
-                break;
-           case SV_LEFT_DPI_DEC:
-                decrease_left_dpi();
-                break;
-           case SV_RIGHT_DPI_INC:
-                increase_right_dpi();
-                break;
-           case SV_RIGHT_DPI_DEC:
-                decrease_right_dpi();
-                break;
-           case SV_LEFT_SCROLL_TOGGLE:
-                global_saved_values.left_scroll = !global_saved_values.left_scroll;
-                write_eeprom_kb();
-                break;
-           case SV_RIGHT_SCROLL_TOGGLE:
-                global_saved_values.right_scroll = !global_saved_values.right_scroll;
-                write_eeprom_kb();
-                break;
-           case SV_RECALIBRATE_POINTER:
-                recalibrate_pointer();
-           default:
-                break;
-        }
-    }
-#endif
-
-  switch (keycode) {
-    default:
-      return true;
-  }
-};
-
-#if defined MH_AUTO_BUTTONS && defined PS2_MOUSE_ENABLE && defined MOUSEKEY_ENABLE
-void ps2_mouse_moved_user(report_mouse_t *mouse_report) {
-  if (mh_auto_buttons_timer) {
-    mh_auto_buttons_timer = timer_read();
-  } else {
-    if (!tp_buttons) {
-      mouse_mode(true);
-  #if defined CONSOLE_ENABLE
-      print("mh_auto_buttons: on\n");
-  #endif
-    }
-  }
-}
-#endif
-
-
-
-#if (defined MH_AUTO_BUTTONS && defined PS2_MOUSE_ENABLE && defined MOUSEKEY_ENABLE) || defined(POINTING_DEVICE_AUTO_MOUSE_MH_ENABLE)
-void matrix_scan_user(void) {
-  if (mh_auto_buttons_timer && (timer_elapsed(mh_auto_buttons_timer) > MH_AUTO_BUTTONS_TIMEOUT)) {
-    if (!tp_buttons) {
-      mouse_mode(false);
-  #if defined CONSOLE_ENABLE
-      print("mh_auto_buttons: off\n");
-  #endif
-    }
-  }
-}
-
-void mouse_mode(bool on) {
-  if (on) {
-    layer_on(MH_AUTO_BUTTONS_LAYER);
-    mh_auto_buttons_timer = timer_read();
-  } else {
-    layer_off(MH_AUTO_BUTTONS_LAYER);
-    mh_auto_buttons_timer = 0;
-  }
-}
-
-#endif // defined MH_AUTO_BUTTONS && defined PS2_MOUSE_ENABLE && #defined MOUSEKEY_ENABLE
