+++ keyboards/svalboard/keymaps/vial/rules.mk
@@ -1,2 +1,3 @@
 VIA_ENABLE = yes
 VIAL_ENABLE = yes
+VIAL_INSECURE = yes
