+++ keyboards/svalboard/keymaps/vial_secure/rules.mk
@@ -1,4 +1,2 @@
 VIA_ENABLE = yes
 VIAL_ENABLE = yes
-VIAL_INSECURE = yes
-
