--- keyboards/svalboard/keymaps/vialblank/rules.mk
@@ -1,4 +1,0 @@
-VIA_ENABLE = yes
-VIAL_ENABLE = yes
-VIAL_INSECURE = yes
-
