+++ keyboards/svalboard/vils/QWERTY Num and Func layers done.vil
