+++ keyboards/svalboard/vils/QWERTY and Num layers done.vil
